Moved the SharedAtoms into a separate file which will eventually get generated, and optimized implementation
I’ve gotten tired of adding shared atoms in 3 places. So I moved the enums, static strings and table into a separate file. Eventually I’ll make a data file with all the strings and generate these files.

I also got rid of the concept of having the shared atoms in a separate atom table. This increased memory usage because all shared atoms would take up room in this table, whether they were used or not. It was also less efficient because you had to search 2 tables. And it was harder to debug because all the non-shared atoms were offset so they were hard to find. Finally, this will make it easier when we redo serialization because we just store a single atom table.

+++ m8rscript/Containers.h
@@ -333,7 +333,7 @@ private:
 template<class T>
 struct CompareKeys
 {
-    int operator()(const T& lhs, const T& rhs) const { return lhs - rhs; }
+    int operator()(const T& lhs, const T& rhs) const { return static_cast<int>(lhs - rhs); }
 };
 
 template<typename Key, typename Value, typename Compare=CompareKeys<Key>>
